+++ PBL3_OnlineShop/PBL3_OnlineShop/Program.cs
@@ -10,6 +10,7 @@ using PBL3_OnlineShop.Services.Admin.Statistic;
 using PBL3_OnlineShop.Services.Admin.User;
 using PBL3_OnlineShop.Services.Cart;
 using PBL3_OnlineShop.Services.Checkout;
+using PBL3_OnlineShop.Services.Product;
 
 namespace PBL3_OnlineShop
 {
@@ -30,12 +31,15 @@ namespace PBL3_OnlineShop
             builder.Services.AddScoped<IStatisticService,StatisticService>();
             builder.Services.AddScoped<IOrderService, OrderService>();
             builder.Services.AddScoped<IUserService, UserService>();
-            builder.Services.AddScoped<IProductService, ProductService>();
+            builder.Services.AddScoped<Services.Admin.Product.IProductService, Services.Admin.Product.ProductService>();
             builder.Services.AddScoped<ICategoryService, CategoryService>();
             builder.Services.AddScoped<ICouponService, CouponService>();
             builder.Services.AddScoped<IAccountService, AccountService>();
             builder.Services.AddScoped<ICartService, CartService>();
             builder.Services.AddScoped<ICheckoutService, CheckoutService>();
+            
+            // Đăng ký dịch vụ ProductService cho phía người dùng
+            builder.Services.AddScoped<Services.Product.IProductService, Services.Product.ProductService>();
 
             builder.Services.AddDistributedMemoryCache(); // Bộ nhớ để lưu session
             builder.Services.AddSession(options =>
